cyber/requirements.txt
fastapi
uvicorn
slowapi
streamlit
plotly
pandas
requests
fpdf2
python-dotenv